moved various components into folders and updated part of css
