+++ frontend/src/AppContextProvider.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import axios from 'axios';
+
+const AppContext = React.createContext({
+    
+});
