+++ frontend/src/components/CodeInput/style.module.css
@@ -9,14 +9,3 @@
     display: flex;
     justify-content: center;
 }
-
-.rootContainer {
-    padding: 15px;
-    position: fixed;
-    left: 0;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    overflow: auto;
-    background: #333333;
-  }
