+++ frontend/src/components/LandingOptions/style.module.css
@@ -1,0 +1,32 @@
+.buttonContainer {
+    margin: 0;
+    position: absolute;
+    height: 300px;
+    width: 70%;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+}
+
+.topButton {
+    background-color: #333333;
+    border-radius: 30px;
+    padding: 20px 40px;
+    width: 100%;
+    font-size: 18px;
+    color: white;
+    box-shadow: 0 0 8px 0 white;
+}
+
+.bottomButton {
+    background-color: #333333;
+    border-radius: 30px;
+    padding: 20px 40px;
+    width: 100%;
+    font-size: 18px;
+    color: white;
+    box-shadow: 0 0 8px 0 white;
+    position: absolute;
+    bottom: 0;
+    left: 0;
+}
