+++ frontend/src/components/SongSearch/style.module.css
@@ -1,0 +1,5 @@
+.image {
+  height: 64px;
+  width: 64px;
+  float: right;
+}
