+++ frontend/src/components/index.js
@@ -1,0 +1,5 @@
+export { default as CodeInput } from './CodeInput';
+export { default as LandingOptions } from './LandingOptions';
+export { default as Playlist } from './Playlist';
+export { default as SongSearch } from './SongSearch';
+export { default as HostCreateRoom } from './HostCreateRoom';
