--- frontend/src/landingPage/style.module.css
@@ -1,19 +1,0 @@
-.buttonContainer {
-    margin: 0;
-    position: absolute;
-    height: 300px;
-    width: 70%;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-}
-
-.rootContainer {
-    position: fixed;
-    left: 0;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    overflow: auto;
-    background: #333333;
-  }
