+++ frontend/src/pages/JoinPlaylistPage/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+import styles from './style.module.css';
+import { CodeInput } from '../../components';
+
+export default function JoinPlaylistPage() {
+  return (
+    <div className={styles.rootContainer}>
+      <CodeInput />
+    </div>
+  );
+}
