+++ frontend/src/pages/JoinPlaylistPage/style.module.css
@@ -1,0 +1,10 @@
+.rootContainer {
+    padding: 15px;
+    position: fixed;
+    left: 0;
+    top: 0;
+    bottom: 0;
+    right: 0;
+    overflow: auto;
+    background: #333333;
+  }
