+++ frontend/src/pages/LandingPage/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import styles from "./style.module.css";
+import { LandingOptions } from "../../components";
+
+export default function LandingPage() {
+
+    return (
+        <div className={styles.rootContainer} >
+            <LandingOptions />
+        </div>
+    );
+}
