+++ frontend/src/pages/LandingPage/style.module.css
@@ -1,0 +1,9 @@
+.rootContainer {
+    position: fixed;
+    left: 0;
+    top: 0;
+    bottom: 0;
+    right: 0;
+    overflow: auto;
+    background: #333333;
+  }
