+++ frontend/src/pages/index.js
@@ -1,0 +1,4 @@
+export { default as HomePage } from './HomePage';
+export { default as HostPage } from './HostPage';
+export { default as JoinPlaylistPage } from './JoinPlaylistPage';
+export { default as LandingPage } from './LandingPage';
